Add separate item table for confirmation page
