+++ pages/hold/request/[id]/edd.tsx
@@ -10,7 +10,7 @@ import {
 import Layout from "../../../../src/components/Layout/Layout"
 import EDDRequestForm from "../../../../src/components/HoldPages/EDDRequestForm"
 import HoldRequestErrorBanner from "../../../../src/components/HoldPages/HoldRequestErrorBanner"
-import HoldItemDetails from "../../../../src/components/HoldPages/HoldItemDetails"
+import HoldRequestItemDetails from "../../../../src/components/HoldPages/HoldRequestItemDetails"
 
 import { SITE_NAME, BASE_URL, PATHS } from "../../../../src/config/constants"
 import useLoading from "../../../../src/hooks/useLoading"
@@ -151,7 +151,7 @@ export default function EDDRequestPage({
         <Heading level="h2" mb="l" size="heading3">
           Request scan
         </Heading>
-        <HoldItemDetails item={item} />
+        <HoldRequestItemDetails item={item} />
         {isLoading || formPosting ? (
           <SkeletonLoader showImage={false} data-testid="edd-request-loading" />
         ) : errorStatus !== "eddUnavailable" ? (
