+++ pages/hold/request/[id]/index.tsx
@@ -11,7 +11,7 @@ import Layout from "../../../../src/components/Layout/Layout"
 
 import HoldRequestForm from "../../../../src/components/HoldPages/HoldRequestForm"
 import HoldRequestErrorBanner from "../../../../src/components/HoldPages/HoldRequestErrorBanner"
-import HoldItemDetails from "../../../../src/components/HoldPages/HoldItemDetails"
+import HoldRequestItemDetails from "../../../../src/components/HoldPages/HoldRequestItemDetails"
 
 import { SITE_NAME, BASE_URL, PATHS } from "../../../../src/config/constants"
 import useLoading from "../../../../src/hooks/useLoading"
@@ -165,7 +165,7 @@ export default function HoldRequestPage({
         <Heading level="h2" mb="l" size="heading3">
           Request for on-site use
         </Heading>
-        <HoldItemDetails item={item} />
+        <HoldRequestItemDetails item={item} />
         {isLoading || formPosting ? (
           <SkeletonLoader
             showImage={false}
