+++ src/components/HoldPages/HoldRequestItemDetails.tsx
@@ -1,10 +1,10 @@
 import { List } from "@nypl/design-system-react-components"
 
-import type Item from "../../../src/models/Item"
+import type Item from "../../models/Item"
 import { LinkedDetailElement, PlainTextElement } from "../BibPage/BibDetail"
 
 import bibDetailStyles from "../../../styles/components/BibDetails.module.scss"
-import { PATHS } from "../../../src/config/constants"
+import { PATHS } from "../../config/constants"
 
 interface HoldItemDetailsProps {
   item: Item
